refactor: Extract `include` transform into its own module
