+++ tools/markdown-magic/src/md-magic.config.cjs
@@ -10,12 +10,10 @@ const scripts = require("markdown-magic-package-scripts");
 
 const { templatesDirectoryPath } = require("./constants.cjs");
 const {
-	formattedEmbeddedContentBody,
 	formattedGeneratedContentBody,
 	formattedSectionText,
 	getPackageMetadata,
 	resolveRelativePackageJsonPath,
-	resolveRelativePath,
 } = require("./utilities.cjs");
 const {
 	apiDocsLinkSectionTransform,
@@ -23,6 +21,7 @@ const {
 	generateInstallationInstructionsSection,
 	generatePackageImportInstructionsSection,
 	generatePackageScriptsSection,
+	includeTransform,
 	installationInstructionsTransform,
 	packageImportInstructionsSectionTransform,
 	packageScriptsSectionTransform,
@@ -208,54 +207,6 @@ function getPackageMetadataFromRelativePath(documentFilePath, packageJsonFilePat
 	return getPackageMetadata(resolvedPackageJsonPath);
 }
 
-/**
- * Embeds contents from the specified file paths within the provided (optional) line boundaries.
- *
- * @param {object} content - The original document file contents.
- * @param {object} options - Transform options.
- * @param {string} options.path - Relative path from the document to the file being embedded.
- * @param {string | undefined} options.start - (optional) First line from the target file to be embedded (inclusive).
- * Expected to be a string-formatted integer.
- * Default: 0.
- * Constraints are the same as those for the `end` parameter to
- * {@link https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/slice#parameters | Array.slice}
- * @param {string | undefined} options.end - (optional) Line of the target file at which to end the embedded range (exclusive).
- * Expected to be a string-formatted integer.
- * Default: <file-line-count> + 1.
- * Constraints are the same as those for the `end` parameter to
- * {@link https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/slice#parameters | Array.slice}
- * @param {object} config - Transform configuration.
- * @param {string} config.originalPath - Path to the document being modified.
- */
-function includeTransform(content, options, config) {
-	const { path: relativeFilePath, start: startLineString, end: endLineString } = options;
-	const { originalPath: documentFilePath } = config;
-
-	const startLine = startLineString === undefined ? undefined : Number.parseInt(startLineString);
-	const endLine = endLineString === undefined ? undefined : Number.parseInt(endLineString);
-
-	if (!relativeFilePath) {
-		throw new Error(
-			"No 'path' parameter provided. Must specify a relative path to the file containing the contents to be embedded.",
-		);
-	}
-
-	const resolvedFilePath = resolveRelativePath(documentFilePath, relativeFilePath);
-
-	try {
-		let fileContents = fs.readFileSync(resolvedFilePath, "utf8");
-		if (startLine || endLine) {
-			const split = fileContents.split(/\r?\n/);
-			fileContents = split.slice(startLine, endLine).join("\n");
-		}
-		const section = formattedSectionText(fileContents.trim());
-
-		return formattedEmbeddedContentBody(section);
-	} catch (error) {
-		console.error(`Exception processing "${resolvedFilePath}":`, error);
-		throw error;
-	}
-}
 
 /**
  * Generates simple README contents for a library package.
